Don't force LF ending in .editorconfig
The change have been made on the yeoman/generator repo too. The line
normalizing when commiting is already activated via .gitattributes.
There's no need to duplicate the normalization in the .editorconfig.
Overall, Windows developper prefer to checkout CRLF and commit LF as this
ensure maximum compatibility and does not constantly break diffing with
git.