Disable brittle tests
These tests fails because with the new way tasks are runned we cannot
expect Mocha to stop the process when the async callback is called.

But on a more high level view, these tests really doesn't exercise anything
useful. This mechanic is already covered by the generator core system
tests. We should rewrite them to be assert behavior specific to yo.